// src/api/generated/model/dirEntry.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Retrowin API
 * Retrowin File Management System API
 * OpenAPI spec version: 0.2.0
 */

export interface DirEntry {
  /** Entry name */
  name: string;
  /** Inode ID */
  inodeId: string;
  /** File type (upper 4 bits of mode) */
  fileType: number;
}